requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
streamlit
plotly
scikit-learn
numpy
db-dtypes
thefuzz
pyarrow